decomposing eigendecomp
