+++ cleverhans/experimental/certification/optimization.py
@@ -9,6 +9,7 @@ import tensorflow as tf
 from tensorflow.contrib import autograph
 from cleverhans.experimental.certification import utils
 from cleverhans.experimental.certification import dual_formulation
+from cleverhans.experimental.certification import eigen
 
 # Bound on lowest value of certificate to check for numerical errors
 LOWER_CERT_BOUND = -10.0
@@ -42,10 +43,12 @@ class Optimization(object):
 
     # The dimensionality of matrix M is the sum of sizes of all layers + 1
     # The + 1 comes due to a row and column of M representing the linear terms
-    self.eig_init_vec_placeholder = tf.placeholder(
-        tf.float32, shape=[1 + self.dual_object.dual_index[-1], 1])
-    self.smooth_placeholder = tf.placeholder(tf.float32, shape=[])
-    self.eig_num_iter_placeholder = tf.placeholder(tf.int32, shape=[])
+    # self.eig_init_vec_placeholder = tf.placeholder(
+    #     tf.float32, shape=[1 + self.dual_object.dual_index[-1], 1])
+    # self.smooth_placeholder = tf.placeholder(tf.float32, shape=[])
+    # self.eig_num_iter_placeholder = tf.placeholder(tf.int32, shape=[])
+    
+    self.eigen = eigen.EigenDecomposition (self.dual_object, self.params['eig_learning_rate'])
     self.sess = sess
 
   def project_dual(self):
@@ -132,75 +135,76 @@ class Optimization(object):
         self.dual_object.input_maxval, self.dual_object.epsilon)
     return projected_dual_object
 
-  def tf_min_eig_vec(self):
-    """Function for min eigen vector using tf's full eigen decomposition."""
-    # Full eigen decomposition requires the explicit psd matrix M
-    _, matrix_m = self.dual_object.get_full_psd_matrix()
-    [eig_vals, eig_vectors] = tf.self_adjoint_eig(matrix_m)
-    index = tf.argmin(eig_vals)
-    return tf.reshape(
-        eig_vectors[:, index], shape=[eig_vectors.shape[0].value, 1])
+#   def tf_min_eig_vec(self):
+#     """Function for min eigen vector using tf's full eigen decomposition."""
+#     # Full eigen decomposition requires the explicit psd matrix M
+#     _, matrix_m = self.dual_object.get_full_psd_matrix()
+#     [eig_vals, eig_vectors] = tf.self_adjoint_eig(matrix_m)
+#     index = tf.argmin(eig_vals)
+#     return tf.reshape(
+#         eig_vectors[:, index], shape=[eig_vectors.shape[0].value, 1])
 
-  def tf_smooth_eig_vec(self):
-    """Function that returns smoothed version of min eigen vector."""
-    _, matrix_m = self.dual_object.get_full_psd_matrix()
-    # Easier to think in terms of max so negating the matrix
-    [eig_vals, eig_vectors] = tf.self_adjoint_eig(-matrix_m)
-    exp_eig_vals = tf.exp(tf.divide(eig_vals, self.smooth_placeholder))
-    scaling_factor = tf.reduce_sum(exp_eig_vals)
-    # Multiplying each eig vector by exponential of corresponding eig value
-    # Scaling factor normalizes the vector to be unit norm
-    eig_vec_smooth = tf.divide(
-        tf.matmul(eig_vectors, tf.diag(tf.sqrt(exp_eig_vals))),
-        tf.sqrt(scaling_factor))
-    return tf.reshape(
-        tf.reduce_sum(eig_vec_smooth, axis=1),
-        shape=[eig_vec_smooth.shape[0].value, 1])
+#   def tf_smooth_eig_vec(self):
+#     """Function that returns smoothed version of min eigen vector."""
+#     _, matrix_m = self.dual_object.get_full_psd_matrix()
+#     # Easier to think in terms of max so negating the matrix
+#     [eig_vals, eig_vectors] = tf.self_adjoint_eig(-matrix_m)
+#     exp_eig_vals = tf.exp(tf.divide(eig_vals, self.smooth_placeholder))
+#     scaling_factor = tf.reduce_sum(exp_eig_vals)
+#     # Multiplying each eig vector by exponential of corresponding eig value
+#     # Scaling factor normalizes the vector to be unit norm
+#     eig_vec_smooth = tf.divide(
+#         tf.matmul(eig_vectors, tf.diag(tf.sqrt(exp_eig_vals))),
+#         tf.sqrt(scaling_factor))
+#     return tf.reshape(
+#         tf.reduce_sum(eig_vec_smooth, axis=1),
+#         shape=[eig_vec_smooth.shape[0].value, 1])
 
-  def get_min_eig_vec_proxy(self, use_tf_eig=False):
-    """Computes the min eigen value and corresponding vector of matrix M.
+#   def get_min_eig_vec_proxy(self, use_tf_eig=False):
+#     """Computes the min eigen value and corresponding vector of matrix M.
 
-    Args:
-      use_tf_eig: Whether to use tf's default full eigen decomposition
+#     Args:
+#       use_tf_eig: Whether to use tf's default full eigen decomposition
 
-    Returns:
-      eig_vec: Minimum absolute eigen value
-      eig_val: Corresponding eigen vector
-    """
-    if use_tf_eig:
-      # If smoothness parameter is too small, essentially no smoothing
-      # Just output the eigen vector corresponding to min
-      return tf.cond(self.smooth_placeholder < 1E-8, self.tf_min_eig_vec,
-                     self.tf_smooth_eig_vec)
+#     Returns:
+#       eig_vec: Minimum absolute eigen value
+#       eig_val: Corresponding eigen vector
+#     """
+#     if use_tf_eig:
+#       # If smoothness parameter is too small, essentially no smoothing
+#       # Just output the eigen vector corresponding to min
+#       return tf.cond(self.smooth_placeholder < 1E-8, self.tf_min_eig_vec,
+#                      self.tf_smooth_eig_vec)
 
-    # Using autograph to automatically handle
-    # the control flow of minimum_eigen_vector
-    min_eigen_tf = autograph.to_graph(utils.minimum_eigen_vector)
+#     # Using autograph to automatically handle
+#     # the control flow of minimum_eigen_vector
+#     min_eigen_tf = autograph.to_graph(utils.minimum_eigen_vector)
 
-    def _vector_prod_fn(x):
-      return self.dual_object.get_psd_product(x)
+#     def _vector_prod_fn(x):
+#       return self.dual_object.get_psd_product(x)
 
-    estimated_eigen_vector = min_eigen_tf(
-        x=self.eig_init_vec_placeholder,
-        num_steps=self.eig_num_iter_placeholder,
-        learning_rate=self.params['eig_learning_rate'],
-        vector_prod_fn=_vector_prod_fn)
-    return estimated_eigen_vector
+#     estimated_eigen_vector = min_eigen_tf(
+#         x=self.eig_init_vec_placeholder,
+#         num_steps=self.eig_num_iter_placeholder,
+#         learning_rate=self.params['eig_learning_rate'],
+#         vector_prod_fn=_vector_prod_fn)
+#     return estimated_eigen_vector
 
   def prepare_for_optimization(self):
     """Create tensorflow op for running one step of descent."""
 
-    self.eig_vec_estimate = self.get_min_eig_vec_proxy()
-    self.stopped_eig_vec_estimate = tf.stop_gradient(self.eig_vec_estimate)
-    # Eig value is v^\top M v, where v is eigen vector
-    self.eig_val_estimate = tf.matmul(
-        tf.transpose(self.stopped_eig_vec_estimate),
-        self.dual_object.get_psd_product(self.stopped_eig_vec_estimate))
+    self.eigen.create_estimates ()
+    # self.eig_vec_estimate = self.get_min_eig_vec_proxy()
+    # self.stopped_eig_vec_estimate = tf.stop_gradient(self.eig_vec_estimate)
+    # # Eig value is v^\top M v, where v is eigen vector
+    # self.eig_val_estimate = tf.matmul(
+    #     tf.transpose(self.stopped_eig_vec_estimate),
+    #     self.dual_object.get_psd_product(self.stopped_eig_vec_estimate))
     # Penalizing negative of min eigen value because we want min eig value
     # to be positive
     self.total_objective = (
         self.dual_object.unconstrained_objective + 0.5 * (tf.square(
-            tf.maximum(-1 * self.penalty_placeholder * self.eig_val_estimate,
+            tf.maximum(-1 * self.penalty_placeholder * self.eigen.eig_val_estimate,
                        0))))
     global_step = tf.Variable(0, trainable=False)
     # Set up learning rate
@@ -271,20 +275,15 @@ class Optimization(object):
         tf.logging.info('Found certificate of robustness!')
         return True
     # Running step
-    step_feed_dict = {
-        self.eig_init_vec_placeholder: eig_init_vec_val,
-        self.eig_num_iter_placeholder: eig_num_iter_val,
-        self.smooth_placeholder: smooth_val,
-        self.penalty_placeholder: penalty_val
-    }
+    step_feed_dict = self.eigen.create_feed_dict(eig_init_vec_val, eig_num_iter_val, smooth_val, penalty_val)
 
     [
-        _, self.current_total_objective, self.current_unconstrained_objective,
+        _, self.current_total_objective, current_unconstrained_objective,
         self.current_eig_vec_val, self.current_eig_val_estimate
     ] = self.sess.run([
         self.opt_one_step, self.total_objective,
-        self.dual_object.unconstrained_objective, self.eig_vec_estimate,
-        self.eig_val_estimate
+        self.dual_object.unconstrained_objective, self.eigen.eig_vec_estimate,
+        self.eigen.eig_val_estimate
     ],
                       feed_dict=step_feed_dict)
 
@@ -293,7 +292,7 @@ class Optimization(object):
           'total_objective':
               float(self.current_total_objective),
           'unconstrained_objective':
-              float(self.current_unconstrained_objective),
+              float(current_unconstrained_objective),
           'min_eig_val_estimate':
               float(self.current_eig_val_estimate)
       }
